Add multi-file benchmark and module caching.
This commit adds the `pathtrace` compilation benchmark, which is based on the NVIDIA nvpro_core path tracing shaders.
It simulates a fairly typical medium-sized codebase, with 17 files, 162 KB, and 4799 lines of GLSL code.
I've ported this line-by-line to Slang code, simulating what I figure a typical developer would do -- turning GLSL
headers into Slang modules, using `import` instead of `#include`, but not changing the structure of the codebase
too much or using too many advanced language features.

This also adds a module caching system to the Slang compiler helper. When the Slang compiler attempts to load a
file, the compiler helper can pre-compile it into a binary module and reuse that for future compilations.